Document the arm-fpu-fix rebase in 5.9.2-1
- arm-fpu-fix patch: Drop the host tools hunk added in 5.9.0-2, fixed upstream

Also rename the patch file from
qtwebengine-opensource-src-5.9.0-arm-fpu-fix.patch to
qtwebengine-opensource-src-5.9.2-arm-fpu-fix.patch because the patch was
rebased. Otherwise, it is easy to miss that the patch was rebased. (I
did, I only noticed it now while working on the RPM Fusion package. And
I might easily have missed it again and run into a build failure due to
that.)