no-sse2 patch: Fix WebRTC aec3 module FTBFS
The aec3 module attempts to do SSE2 runtime detection in a way that does
not compile with GCC, so disable it. (SSE2 use in that module is now
compile-time only, i.e., only enabled for the x86_64 package in Fedora.)

The patch also optimizes x86_64 by making the SSE2 detection return a
compile-time constant, which upstream neglected doing.