Drop obsolete no-icudtl-dat patch, code has been fixed upstream since 5.11.0
* Wed Jul 20 2022 Kevin Kofler <Kevin@tigcc.ticalc.org> - 5.15.9-3
- Drop obsolete no-icudtl-dat patch, code has been fixed upstream since 5.11.0

If you look closely at the patch, you will see that the code that is
removed does not actually mention the offending icudtl.dat anywhere.
This upstream code snippet was fixed back in 5.11.0 to check for
qtwebengine_resources.pak instead, which is always installed. (Compare
my original patch with the version rebased by Rex Dieter for 5.11.0.)

Searching for alternate resource paths is still pointless in packaged
QtWebEngine, mind you. (The file installed by the package should always
be found.) So the patch did not hurt. But still, it is no longer
necessary to patch this out.